Fixes, Updates
Fixed a big with dynamic resource linking that caused a crash in
presentation foundation when changing the current view from a view which
contained a rectangle with the opacity mask set to a dynamic resource of
appbar_WHATEVER. Not sure if this is a bug in my application, mahapps
etc so no bug report made. Fix is to use `StaticResource` instead.

Cleaned up `BaseView` to move stuff that should have been inside
`IViewController` where it should be.

From a design perspective, added/finished the design of the
`Settings.xaml` view. Classes for model/view model are created but no
logic yet implemented.