Fixes, Additions
Removed the old system in `FilteringCategory` for automatically creating
unique category IDs. The old system used interlocked ops on a private
static int. This wasn't logically sound and or sufficient for the
intended use, especially when you consider recycling these numbers at
random when the user deletes a list. This system was replaced with a
`ConcurrentBag` which holds all possible filtering list values. Taken
values are pushed back to the bag when `FilterCategory` objects are
disposed.

All filter categories created at held at the App.xaml.cs level and
supplied via an observable collection to views that must be aware of
these categories (stats, settings etc).

Other than that, reorganizing, filling out stubs and additions of
convenience stuff such as the ability for all `BaseView` objects to
raise alerts to the user easily.